Add a separate repo variable file
