+++ app/client/cypress/integration/Regression_TestSuite/ClientSideTests/AuditLogs/Audit_logs_spec.js
@@ -1,4 +1,4 @@
-import { REPO, REPO_VARIABLES } from "../../../../fixtures/variables";
+import { REPO, CURRENT_REPO } from "../../../../fixtures/REPO";
 
 const Access = {
   AdminSettingsEntryLink: ".admin-settings-menu-option",
@@ -29,7 +29,7 @@ const locators = { ...Access, ...UpgradePage };
 
 describe("Audit logs", () => {
   it("1. Super user can access audit logs page", () => {
-    if (REPO_VARIABLES.Edition === REPO.CE) {
+    if (CURRENT_REPO === REPO.CE) {
       cy.LogOut();
       cy.LoginFromAPI(Cypress.env("USERNAME"), Cypress.env("PASSWORD"));
       cy.visit("/applications");
