+++ app/client/cypress/integration/Regression_TestSuite/ClientSideTests/Git/GitSync/RepoLimitExceededErrorModal_spec.js
@@ -1,6 +1,6 @@
 import gitSyncLocators from "../../../../../locators/gitSyncLocators";
 import * as _ from "../../../../../support/Objects/ObjectsCore";
-import { REPO, REPO_VARIABLES } from "../../../../../fixtures/variables";
+import { REPO, CURRENT_REPO } from "../../../../../fixtures/variables";
 
 let repoName1, repoName2, repoName3, repoName4, windowOpenSpy;
 describe("Repo Limit Exceeded Error Modal", function() {
@@ -47,7 +47,7 @@ describe("Repo Limit Exceeded Error Modal", function() {
     // cy.createAppAndConnectGit(repoName3, false);
     // cy.createAppAndConnectGit(repoName4, false, true);
 
-    if (REPO_VARIABLES.Edition === REPO.CE) {
+    if (CURRENT_REPO === REPO.CE) {
       cy.get(gitSyncLocators.repoLimitExceededErrorModal).should("exist");
 
       // title and info text checking
